Clenshaw-Curtis Quadrature Formula

Clenshaw-Curtis Quadrature Formula

clenshaw_curtis(f, a = -1, b = 1, n = 1024, ...)

Arguments

f
function, the integrand, without singularities.

a, b
lower and upper limit of the integral; must be finite.

n
Number of Chebyshev nodes to account for.

…
Additional parameters to be passed to the function

Details

Clenshaw-Curtis quadrature is based on sampling the integrand on Chebyshev points, an operation that can be implemented using the Fast Fourier Transform.

Value

Numerical scalar, the value of the integral.

References

Trefethen, L. N. (2008). Is Gauss Quadrature Better Than Clenshaw-Curtis? SIAM Review, Vol. 50, No. 1, pp 67--87.

See also

gaussLegendre, gauss_kronrod

Examples

##  Quadrature with Chebyshev nodes and weights
f <- function(x) sin(x+cos(10*exp(x))/3)
# NOT RUN {
ezplot(f, -1, 1, fill = TRUE)
# }
cc <- clenshaw_curtis(f, n = 64)  #=>  0.0325036517151 , true error > 1.3e-10