compan {pracma}	R Documentation

Companion Matrix

Description

Computes the companion matrix of a real or complex vector.

Usage

  compan(p)

Arguments

p
vector representing a polynomial

Details

Computes the companion matrix corresponding to the vector p with -p[2:length(p)]/p[1] as first row.

The eigenvalues of this matrix are the roots of the polynomial.

Value

A square matrix of length(p)-1 rows and columns

See Also

roots

Examples

  p <- c(1, 0, -7, 6)
  compan(p)
  # 0  7 -6
  # 1  0  0
  # 0  1  0

[Package pracma version 2.3.8 Index]