Vector Cross Product

Vector or cross product

cross(x, y)

Arguments

x
numeric vector or matrix

y
numeric vector or matrix

Details

Computes the cross (or: vector) product of vectors in 3 dimensions. In case of matrices it takes the first dimension of length 3 and computes the cross product between corresponding columns or rows.

The more general cross product of n-1 vectors in n-dimensional space is realized as crossn.

Value

3-dim. vector if x and < are vectors, a matrix of 3-dim. vectors if x and y are matrices themselves.

See also

Examples

  cross(c(1, 2, 3), c(4, 5, 6))  # -3  6 -3

#> [1] -3  6 -3