dot {pracma}	R Documentation

Scalar Product

Description

'dot' or 'scalar' product of vectors or pairwise columns of matrices.

Usage

  dot(x, y)

Arguments

x
numeric vector or matrix

y
numeric vector or matrix

Details

Returns the 'dot' or 'scalar' product of vectors or columns of matrices. Two vectors must be of same length, two matrices must be of the same size. If x and y are column or row vectors, their dot product will be computed as if they were simple vectors.

Value

A scalar or vector of length the number of columns of x and y.

See Also

cross

Examples

  dot(1:5, 1:5)  #=> 55
  # Length of space diagonal in 3-dim- cube:
  sqrt(dot(c(1,1,1), c(1,1,1)))  #=> 1.732051

[Package pracma version 2.3.8 Index]