eig {pracma}	R Documentation

Eigenvalue Function (Matlab Style)

Description

Eigenvalues of a matrix

Usage

  eig(a)

Arguments

a
real or complex square matrix

Details

Computes the eigenvalues of a square matrix of real or complex numbers, using the R routine eigen without computing the eigenvectors.

Value

Vector of eigenvalues

See Also

compan

Examples

  eig(matrix(c(1,-1,-1,1), 2, 2))   #=> 2 0
  eig(matrix(c(1,1,-1,1), 2, 2))    # complex values
  eig(matrix(c(0,1i,-1i,0), 2, 2))  # real values

[Package pracma version 2.3.8 Index]