hilb {pracma}	R Documentation

Hilbert Matrix

Description

Generate Hilbert matrix of dimension n

Usage

  hilb(n)

Arguments

n
positive integer specifying the dimension of the Hilbert matrix

Details

Generate the Hilbert matrix H of dimension n with elements H[i, j] = 1/(i+j-1).

(Note: This matrix is ill-conditioned, see e.g. det(hilb(6)).)

Value

matrix of dimension n

See Also

vander

Examples

  hilb(5)

[Package pracma version 2.3.8 Index]