lebesgue {pracma}	R Documentation

Lebesgue Constant

Description

Estimates the Lebesgue constant.

Usage

lebesgue(x, refine = 4, plotting = FALSE)

Arguments

x
numeric vector of grid points

refine
refine the grid with 2^refine grid points; can only be an integer between 2 and 10, default 4.

plotting
shall the Lebesgue function be plotted.

Details

The Lebesgue constant gives an estimation ||P_n f|| ≤ L ||f|| (in minimax norm) where P_n f is the interpolating polynomial of order n for f on an interval [a, b].

Value

Lebesgue constant for the given grid points.

Note

The Lebesgue constant plays an important role when estimating the distance of interpolating polynomials from the minimax solution (see the Remez algorithm).

References

Berrut, J.-P., and L. Nick Trefethen (2004). “Barycentric Lagrange Interpolation”. SIAM Review, Vol. 46(3), pp.501–517.

See Also

barylag

Examples

lebesgue(seq(0, 1, length.out = 6))  #=> 3.100425

[Package pracma version 2.3.8 Index]