meshgrid {pracma}	R Documentation

Generate a Mesh Grid

Description

Generate two matrices for use in three-dimensional plots.

Usage

meshgrid(x, y = x)

Arguments

x
numerical vector, represents points along the x-axis.

y
numerical vector, represents points along the y-axis.

Details

The rows of the output array X are copies of the vector x; columns of the output array Y are copies of the vector y.

Value

Returns two matrices as a list with X and Y components.

Note

The three-dimensional variant meshgrid(x, y, z) is not yet implemented.

See Also

outer

Examples

meshgrid(1:5)$X
meshgrid(c(1, 2, 3), c(11, 12))

[Package pracma version 2.3.8 Index]