Mode {pracma}	R Documentation

Mode function (Matlab style)

Description

Most frequent value in vector or matrix

Usage

Mode(x)

Arguments

x
Real or complex vector or of factor levels.

Details

Computes the ‘sample mode’, i.e. the most frequently occurring value in x.

Among values occurring equally frequently, Mode() chooses the smallest one (for a numeric vector), one with a smallest absolute value (for complex ones) or the first occurring value (for factor levels).

A matrix will be changed to a vector.

Value

One element from x and of the same type. The number of occurrences will not be returned.

Note

In Matlab/Octave an array dimension can be selected along which to find the mode value; this has not been realized here.

Shadows the R function mode that returns essentially the type of an object.

See Also

median

Examples

x <- round(rnorm(1000), 2)
Mode(x)

[Package pracma version 2.3.8 Index]