moler {pracma}	R Documentation

Moler Matrix

Description

Generate the Moler matrix of size n x n. The Moler matrix is for testing eigenvalue computations.

Usage

moler(n)

Arguments

n
integer

Details

The Moler matrix for testing eigenvalue computations is a symmetric matrix with exactly one small eigenvalue.

Value

matrix of size n x n

See Also

wilkinson

Examples

(a <- moler(10))
min(eig(a))

[Package pracma version 2.3.8 Index]