Orthogonal Distance Regression

Orthogonal Distance Regression (ODR, a.k.a. total least squares) is a regression technique in which observational errors on both dependent and independent variables are taken into account.

odregress(x, y)

Arguments

x
matrix of independent variables.

y
vector representing dependent variable.

Details

The implementation used here is applying PCA resp. the singular value decomposition on the matrix of independent and dependent variables.

Value

Returns list with components coeff linear coefficients and intercept term, ssq sum of squares of orthogonal distances to the linear line or hyperplane, err the orthogonal distances, fitted the fitted values, resid the residuals, and normal the normal vector to the hyperplane.

References

Golub, G.H., and C.F. Van Loan (1980). An analysis of the total least squares problem. Numerical Analysis, Vol. 17, pp. 883-893. http://www.cs.cornell.edu/cv/ResearchPDF/Analysis.total.least.squares.prob.pdf

See ODRPACK or ODRPACK95 (TOMS Algorithm 676). URL: http://docs.scipy.org/doc/external/odr_ams.pdf URL: http://semi.vt.edu/presentations/SEMI-March05_Watson.pdf

Note

The ``geometric mean" regression not implemented because questionable.

See also

lm

Examples

# Example in one dimension
x <- c(1.0, 0.6, 1.2, 1.4, 0.2)
y <- c(0.5, 0.3, 0.7, 1.0, 0.2)
odr <- odregress(x, y)
( cc <- odr$coeff )

#> [1]  0.65145762 -0.03328271

# [1]  0.65145762 -0.03328271
lm(y ~ x)

#> 
#> Call:
#> lm(formula = y ~ x)
#> 
#> Coefficients:
#> (Intercept)            x  
#>    -0.01379      0.62931  
#> 

# Coefficients:
# (Intercept)            x 
#    -0.01379      0.62931 

# Prediction
xnew <- seq(0, 1.5, by = 0.25)
( ynew <- cbind(xnew, 1) %*% cc )

#>             [,1]
#> [1,] -0.03328271
#> [2,]  0.12958170
#> [3,]  0.29244610
#> [4,]  0.45531051
#> [5,]  0.61817491
#> [6,]  0.78103932
#> [7,]  0.94390373

# NOT RUN {
plot(x, y, xlim=c(0, 1.5), ylim=c(0, 1.2), main="Orthogonal Regression")
abline(lm(y ~ x), col="blue")
lines(c(0, 1.5), cc[1]*c(0, 1.5) + cc[2], col="red")
points(xnew, ynew, col = "red")
grid()
# }

# Example in two dimensions
x <- cbind(c(0.92, 0.89, 0.85, 0.05, 0.62, 0.55, 0.02, 0.73, 0.77, 0.57),
           c(0.66, 0.47, 0.40, 0.23, 0.17, 0.09, 0.92, 0.06, 0.09, 0.60))
y <- x %*% c(0.5, 1.5) + 1
odr <- odregress(x, y); odr

#> $coeff
#> [1] 0.5 1.5 1.0
#> 
#> $ssq
#> [1] 1.124743e-31
#> 
#> $err
#>               [,1]
#>  [1,] 5.551115e-17
#>  [2,] 1.110223e-16
#>  [3,] 8.326673e-17
#>  [4,] 0.000000e+00
#>  [5,] 8.326673e-17
#>  [6,] 5.551115e-17
#>  [7,] 0.000000e+00
#>  [8,] 0.000000e+00
#>  [9,] 5.551115e-17
#> [10,] 2.775558e-16
#> 
#> $fitted
#>        [,1]
#>  [1,] 2.450
#>  [2,] 2.150
#>  [3,] 2.025
#>  [4,] 1.370
#>  [5,] 1.565
#>  [6,] 1.410
#>  [7,] 2.390
#>  [8,] 1.455
#>  [9,] 1.520
#> [10,] 2.185
#> 
#> $resid
#>               [,1]
#>  [1,] 0.000000e+00
#>  [2,] 0.000000e+00
#>  [3,] 0.000000e+00
#>  [4,] 2.220446e-16
#>  [5,] 0.000000e+00
#>  [6,] 2.220446e-16
#>  [7,] 4.440892e-16
#>  [8,] 0.000000e+00
#>  [9,] 0.000000e+00
#> [10,] 0.000000e+00
#> 
#> $normal
#> [1]  0.2672612  0.8017837 -0.5345225
#> 

# $coeff
# [1] 0.5 1.5 1.0
# $ssq
# [1] 1.473336e-31

y <- y + rep(c(0.1, -0.1), 5)
odr <- odregress(x, y); odr

#> $coeff
#> [1] 0.5921823 1.6750269 0.8803822
#> 
#> $ssq
#> [1] 0.02168174
#> 
#> $err
#>              [,1]
#>  [1,] 0.009462925
#>  [2,] 0.070969459
#>  [3,] 0.034949336
#>  [4,] 0.012383994
#>  [5,] 0.065094720
#>  [6,] 0.022972669
#>  [7,] 0.027835729
#>  [8,] 0.028535956
#>  [9,] 0.065180464
#> [10,] 0.067661096
#> 
#> $fitted
#>           [,1]
#>  [1,] 2.530708
#>  [2,] 2.194687
#>  [3,] 2.053748
#>  [4,] 1.295248
#>  [5,] 1.532290
#>  [6,] 1.356835
#>  [7,] 2.433251
#>  [8,] 1.413177
#>  [9,] 1.487115
#> [10,] 2.222942
#> 
#> $resid
#>              [,1]
#>  [1,]  0.01929229
#>  [2,] -0.14468713
#>  [3,]  0.07125205
#>  [4,] -0.02524754
#>  [5,]  0.13271016
#>  [6,] -0.04683493
#>  [7,]  0.05674936
#>  [8,] -0.05817693
#>  [9,]  0.13288497
#> [10,] -0.13794229
#> 
#> $normal
#> [1]  0.2904671  0.8216056 -0.4905029
#> 

# $coeff
# [1] 0.5921823 1.6750269 0.8803822
# $ssq
# [1] 0.02168174

lm(y ~ x)

#> 
#> Call:
#> lm(formula = y ~ x)
#> 
#> Coefficients:
#> (Intercept)           x1           x2  
#>      0.9153       0.5671       1.6209  
#> 

# Coefficients:
# (Intercept)           x1           x2  
#      0.9153       0.5671       1.6209