Range Space

Range space or image of a matrix.

orth(M)

Arguments

M
Numeric matrix; vectors will be considered as column vectors.

Details

B=orth(A) returns an orthonormal basis for the range of A. The columns of B span the same space as the columns of A, and the columns of B are orthogonal to each other.

The number of columns of B is the rank of A.

Value

Matrix of orthogonal columns, spanning the image of M.

References

Trefethen, L. N., and D. Bau III. (1997). Numerical Linear Algebra. SIAM, Philadelphia.

See also

nullspace

Examples

M <- matrix(1:12, 3, 4)
Rank(M)                 #=> 2

#> [1] 2

orth(M)

#>            [,1]        [,2]
#> [1,] -0.5045331 -0.76077568
#> [2,] -0.5745157 -0.05714052
#> [3,] -0.6444983  0.64649464