Generate Permutations

Generates all permutations of a vector a.

perms(a)

Arguments

a
numeric vector of some length n

Details

If a is a vector of length n, generate all permutations of the elements in a as a matrix of size n! x n where each row represents one permutation.

A matrix will be expanded as vector.

Value

matrix of permutations of the elements of a

Note

Not feasible for length(a) > 10.

See also

Examples

perms(6)

#>      [,1]
#> [1,]    6

perms(1:6)

#>        [,1] [,2] [,3] [,4] [,5] [,6]
#>   [1,]    6    5    4    3    2    1
#>   [2,]    6    5    4    3    1    2
#>   [3,]    6    5    4    2    3    1
#>   [4,]    6    5    4    2    1    3
#>   [5,]    6    5    4    1    2    3
#>   [6,]    6    5    4    1    3    2
#>   [7,]    6    5    3    4    2    1
#>   [8,]    6    5    3    4    1    2
#>   [9,]    6    5    3    2    4    1
#>  [10,]    6    5    3    2    1    4
#>  [11,]    6    5    3    1    2    4
#>  [12,]    6    5    3    1    4    2
#>  [13,]    6    5    2    3    4    1
#>  [14,]    6    5    2    3    1    4
#>  [15,]    6    5    2    4    3    1
#>  [16,]    6    5    2    4    1    3
#>  [17,]    6    5    2    1    4    3
#>  [18,]    6    5    2    1    3    4
#>  [19,]    6    5    1    3    2    4
#>  [20,]    6    5    1    3    4    2
#>  [21,]    6    5    1    2    3    4
#>  [22,]    6    5    1    2    4    3
#>  [23,]    6    5    1    4    2    3
#>  [24,]    6    5    1    4    3    2
#>  [25,]    6    4    5    3    2    1
#>  [26,]    6    4    5    3    1    2
#>  [27,]    6    4    5    2    3    1
#>  [28,]    6    4    5    2    1    3
#>  [29,]    6    4    5    1    2    3
#>  [30,]    6    4    5    1    3    2
#>  [31,]    6    4    3    5    2    1
#>  [32,]    6    4    3    5    1    2
#>  [33,]    6    4    3    2    5    1
#>  [34,]    6    4    3    2    1    5
#>  [35,]    6    4    3    1    2    5
#>  [36,]    6    4    3    1    5    2
#>  [37,]    6    4    2    3    5    1
#>  [38,]    6    4    2    3    1    5
#>  [39,]    6    4    2    5    3    1
#>  [40,]    6    4    2    5    1    3
#>  [41,]    6    4    2    1    5    3
#>  [42,]    6    4    2    1    3    5
#>  [43,]    6    4    1    3    2    5
#>  [44,]    6    4    1    3    5    2
#>  [45,]    6    4    1    2    3    5
#>  [46,]    6    4    1    2    5    3
#>  [47,]    6    4    1    5    2    3
#>  [48,]    6    4    1    5    3    2
#>  [49,]    6    3    4    5    2    1
#>  [50,]    6    3    4    5    1    2
#>  [51,]    6    3    4    2    5    1
#>  [52,]    6    3    4    2    1    5
#>  [53,]    6    3    4    1    2    5
#>  [54,]    6    3    4    1    5    2
#>  [55,]    6    3    5    4    2    1
#>  [56,]    6    3    5    4    1    2
#>  [57,]    6    3    5    2    4    1
#>  [58,]    6    3    5    2    1    4
#>  [59,]    6    3    5    1    2    4
#>  [60,]    6    3    5    1    4    2
#>  [61,]    6    3    2    5    4    1
#>  [62,]    6    3    2    5    1    4
#>  [63,]    6    3    2    4    5    1
#>  [64,]    6    3    2    4    1    5
#>  [65,]    6    3    2    1    4    5
#>  [66,]    6    3    2    1    5    4
#>  [67,]    6    3    1    5    2    4
#>  [68,]    6    3    1    5    4    2
#>  [69,]    6    3    1    2    5    4
#>  [70,]    6    3    1    2    4    5
#>  [71,]    6    3    1    4    2    5
#>  [72,]    6    3    1    4    5    2
#>  [73,]    6    2    4    3    5    1
#>  [74,]    6    2    4    3    1    5
#>  [75,]    6    2    4    5    3    1
#>  [76,]    6    2    4    5    1    3
#>  [77,]    6    2    4    1    5    3
#>  [78,]    6    2    4    1    3    5
#>  [79,]    6    2    3    4    5    1
#>  [80,]    6    2    3    4    1    5
#>  [81,]    6    2    3    5    4    1
#>  [82,]    6    2    3    5    1    4
#>  [83,]    6    2    3    1    5    4
#>  [84,]    6    2    3    1    4    5
#>  [85,]    6    2    5    3    4    1
#>  [86,]    6    2    5    3    1    4
#>  [87,]    6    2    5    4    3    1
#>  [88,]    6    2    5    4    1    3
#>  [89,]    6    2    5    1    4    3
#>  [90,]    6    2    5    1    3    4
#>  [91,]    6    2    1    3    5    4
#>  [92,]    6    2    1    3    4    5
#>  [93,]    6    2    1    5    3    4
#>  [94,]    6    2    1    5    4    3
#>  [95,]    6    2    1    4    5    3
#>  [96,]    6    2    1    4    3    5
#>  [97,]    6    1    4    3    2    5
#>  [98,]    6    1    4    3    5    2
#>  [99,]    6    1    4    2    3    5
#> [100,]    6    1    4    2    5    3
#> [101,]    6    1    4    5    2    3
#> [102,]    6    1    4    5    3    2
#> [103,]    6    1    3    4    2    5
#> [104,]    6    1    3    4    5    2
#> [105,]    6    1    3    2    4    5
#> [106,]    6    1    3    2    5    4
#> [107,]    6    1    3    5    2    4
#> [108,]    6    1    3    5    4    2
#> [109,]    6    1    2    3    4    5
#> [110,]    6    1    2    3    5    4
#> [111,]    6    1    2    4    3    5
#> [112,]    6    1    2    4    5    3
#> [113,]    6    1    2    5    4    3
#> [114,]    6    1    2    5    3    4
#> [115,]    6    1    5    3    2    4
#> [116,]    6    1    5    3    4    2
#> [117,]    6    1    5    2    3    4
#> [118,]    6    1    5    2    4    3
#> [119,]    6    1    5    4    2    3
#> [120,]    6    1    5    4    3    2
#> [121,]    5    6    4    3    2    1
#> [122,]    5    6    4    3    1    2
#> [123,]    5    6    4    2    3    1
#> [124,]    5    6    4    2    1    3
#> [125,]    5    6    4    1    2    3
#> [126,]    5    6    4    1    3    2
#> [127,]    5    6    3    4    2    1
#> [128,]    5    6    3    4    1    2
#> [129,]    5    6    3    2    4    1
#> [130,]    5    6    3    2    1    4
#> [131,]    5    6    3    1    2    4
#> [132,]    5    6    3    1    4    2
#> [133,]    5    6    2    3    4    1
#> [134,]    5    6    2    3    1    4
#> [135,]    5    6    2    4    3    1
#> [136,]    5    6    2    4    1    3
#> [137,]    5    6    2    1    4    3
#> [138,]    5    6    2    1    3    4
#> [139,]    5    6    1    3    2    4
#> [140,]    5    6    1    3    4    2
#> [141,]    5    6    1    2    3    4
#> [142,]    5    6    1    2    4    3
#> [143,]    5    6    1    4    2    3
#> [144,]    5    6    1    4    3    2
#> [145,]    5    4    6    3    2    1
#> [146,]    5    4    6    3    1    2
#> [147,]    5    4    6    2    3    1
#> [148,]    5    4    6    2    1    3
#> [149,]    5    4    6    1    2    3
#> [150,]    5    4    6    1    3    2
#> [151,]    5    4    3    6    2    1
#> [152,]    5    4    3    6    1    2
#> [153,]    5    4    3    2    6    1
#> [154,]    5    4    3    2    1    6
#> [155,]    5    4    3    1    2    6
#> [156,]    5    4    3    1    6    2
#> [157,]    5    4    2    3    6    1
#> [158,]    5    4    2    3    1    6
#> [159,]    5    4    2    6    3    1
#> [160,]    5    4    2    6    1    3
#> [161,]    5    4    2    1    6    3
#> [162,]    5    4    2    1    3    6
#> [163,]    5    4    1    3    2    6
#> [164,]    5    4    1    3    6    2
#> [165,]    5    4    1    2    3    6
#> [166,]    5    4    1    2    6    3
#> [167,]    5    4    1    6    2    3
#> [168,]    5    4    1    6    3    2
#> [169,]    5    3    4    6    2    1
#> [170,]    5    3    4    6    1    2
#> [171,]    5    3    4    2    6    1
#> [172,]    5    3    4    2    1    6
#> [173,]    5    3    4    1    2    6
#> [174,]    5    3    4    1    6    2
#> [175,]    5    3    6    4    2    1
#> [176,]    5    3    6    4    1    2
#> [177,]    5    3    6    2    4    1
#> [178,]    5    3    6    2    1    4
#> [179,]    5    3    6    1    2    4
#> [180,]    5    3    6    1    4    2
#> [181,]    5    3    2    6    4    1
#> [182,]    5    3    2    6    1    4
#> [183,]    5    3    2    4    6    1
#> [184,]    5    3    2    4    1    6
#> [185,]    5    3    2    1    4    6
#> [186,]    5    3    2    1    6    4
#> [187,]    5    3    1    6    2    4
#> [188,]    5    3    1    6    4    2
#> [189,]    5    3    1    2    6    4
#> [190,]    5    3    1    2    4    6
#> [191,]    5    3    1    4    2    6
#> [192,]    5    3    1    4    6    2
#> [193,]    5    2    4    3    6    1
#> [194,]    5    2    4    3    1    6
#> [195,]    5    2    4    6    3    1
#> [196,]    5    2    4    6    1    3
#> [197,]    5    2    4    1    6    3
#> [198,]    5    2    4    1    3    6
#> [199,]    5    2    3    4    6    1
#> [200,]    5    2    3    4    1    6
#> [201,]    5    2    3    6    4    1
#> [202,]    5    2    3    6    1    4
#> [203,]    5    2    3    1    6    4
#> [204,]    5    2    3    1    4    6
#> [205,]    5    2    6    3    4    1
#> [206,]    5    2    6    3    1    4
#> [207,]    5    2    6    4    3    1
#> [208,]    5    2    6    4    1    3
#> [209,]    5    2    6    1    4    3
#> [210,]    5    2    6    1    3    4
#> [211,]    5    2    1    3    6    4
#> [212,]    5    2    1    3    4    6
#> [213,]    5    2    1    6    3    4
#> [214,]    5    2    1    6    4    3
#> [215,]    5    2    1    4    6    3
#> [216,]    5    2    1    4    3    6
#> [217,]    5    1    4    3    2    6
#> [218,]    5    1    4    3    6    2
#> [219,]    5    1    4    2    3    6
#> [220,]    5    1    4    2    6    3
#> [221,]    5    1    4    6    2    3
#> [222,]    5    1    4    6    3    2
#> [223,]    5    1    3    4    2    6
#> [224,]    5    1    3    4    6    2
#> [225,]    5    1    3    2    4    6
#> [226,]    5    1    3    2    6    4
#> [227,]    5    1    3    6    2    4
#> [228,]    5    1    3    6    4    2
#> [229,]    5    1    2    3    4    6
#> [230,]    5    1    2    3    6    4
#> [231,]    5    1    2    4    3    6
#> [232,]    5    1    2    4    6    3
#> [233,]    5    1    2    6    4    3
#> [234,]    5    1    2    6    3    4
#> [235,]    5    1    6    3    2    4
#> [236,]    5    1    6    3    4    2
#> [237,]    5    1    6    2    3    4
#> [238,]    5    1    6    2    4    3
#> [239,]    5    1    6    4    2    3
#> [240,]    5    1    6    4    3    2
#> [241,]    4    5    6    3    2    1
#> [242,]    4    5    6    3    1    2
#> [243,]    4    5    6    2    3    1
#> [244,]    4    5    6    2    1    3
#> [245,]    4    5    6    1    2    3
#> [246,]    4    5    6    1    3    2
#> [247,]    4    5    3    6    2    1
#> [248,]    4    5    3    6    1    2
#> [249,]    4    5    3    2    6    1
#> [250,]    4    5    3    2    1    6
#> [251,]    4    5    3    1    2    6
#> [252,]    4    5    3    1    6    2
#> [253,]    4    5    2    3    6    1
#> [254,]    4    5    2    3    1    6
#> [255,]    4    5    2    6    3    1
#> [256,]    4    5    2    6    1    3
#> [257,]    4    5    2    1    6    3
#> [258,]    4    5    2    1    3    6
#> [259,]    4    5    1    3    2    6
#> [260,]    4    5    1    3    6    2
#> [261,]    4    5    1    2    3    6
#> [262,]    4    5    1    2    6    3
#> [263,]    4    5    1    6    2    3
#> [264,]    4    5    1    6    3    2
#> [265,]    4    6    5    3    2    1
#> [266,]    4    6    5    3    1    2
#> [267,]    4    6    5    2    3    1
#> [268,]    4    6    5    2    1    3
#> [269,]    4    6    5    1    2    3
#> [270,]    4    6    5    1    3    2
#> [271,]    4    6    3    5    2    1
#> [272,]    4    6    3    5    1    2
#> [273,]    4    6    3    2    5    1
#> [274,]    4    6    3    2    1    5
#> [275,]    4    6    3    1    2    5
#> [276,]    4    6    3    1    5    2
#> [277,]    4    6    2    3    5    1
#> [278,]    4    6    2    3    1    5
#> [279,]    4    6    2    5    3    1
#> [280,]    4    6    2    5    1    3
#> [281,]    4    6    2    1    5    3
#> [282,]    4    6    2    1    3    5
#> [283,]    4    6    1    3    2    5
#> [284,]    4    6    1    3    5    2
#> [285,]    4    6    1    2    3    5
#> [286,]    4    6    1    2    5    3
#> [287,]    4    6    1    5    2    3
#> [288,]    4    6    1    5    3    2
#> [289,]    4    3    6    5    2    1
#> [290,]    4    3    6    5    1    2
#> [291,]    4    3    6    2    5    1
#> [292,]    4    3    6    2    1    5
#> [293,]    4    3    6    1    2    5
#> [294,]    4    3    6    1    5    2
#> [295,]    4    3    5    6    2    1
#> [296,]    4    3    5    6    1    2
#> [297,]    4    3    5    2    6    1
#> [298,]    4    3    5    2    1    6
#> [299,]    4    3    5    1    2    6
#> [300,]    4    3    5    1    6    2
#> [301,]    4    3    2    5    6    1
#> [302,]    4    3    2    5    1    6
#> [303,]    4    3    2    6    5    1
#> [304,]    4    3    2    6    1    5
#> [305,]    4    3    2    1    6    5
#> [306,]    4    3    2    1    5    6
#> [307,]    4    3    1    5    2    6
#> [308,]    4    3    1    5    6    2
#> [309,]    4    3    1    2    5    6
#> [310,]    4    3    1    2    6    5
#> [311,]    4    3    1    6    2    5
#> [312,]    4    3    1    6    5    2
#> [313,]    4    2    6    3    5    1
#> [314,]    4    2    6    3    1    5
#> [315,]    4    2    6    5    3    1
#> [316,]    4    2    6    5    1    3
#> [317,]    4    2    6    1    5    3
#> [318,]    4    2    6    1    3    5
#> [319,]    4    2    3    6    5    1
#> [320,]    4    2    3    6    1    5
#> [321,]    4    2    3    5    6    1
#> [322,]    4    2    3    5    1    6
#> [323,]    4    2    3    1    5    6
#> [324,]    4    2    3    1    6    5
#> [325,]    4    2    5    3    6    1
#> [326,]    4    2    5    3    1    6
#> [327,]    4    2    5    6    3    1
#> [328,]    4    2    5    6    1    3
#> [329,]    4    2    5    1    6    3
#> [330,]    4    2    5    1    3    6
#> [331,]    4    2    1    3    5    6
#> [332,]    4    2    1    3    6    5
#> [333,]    4    2    1    5    3    6
#> [334,]    4    2    1    5    6    3
#> [335,]    4    2    1    6    5    3
#> [336,]    4    2    1    6    3    5
#> [337,]    4    1    6    3    2    5
#> [338,]    4    1    6    3    5    2
#> [339,]    4    1    6    2    3    5
#> [340,]    4    1    6    2    5    3
#> [341,]    4    1    6    5    2    3
#> [342,]    4    1    6    5    3    2
#> [343,]    4    1    3    6    2    5
#> [344,]    4    1    3    6    5    2
#> [345,]    4    1    3    2    6    5
#> [346,]    4    1    3    2    5    6
#> [347,]    4    1    3    5    2    6
#> [348,]    4    1    3    5    6    2
#> [349,]    4    1    2    3    6    5
#> [350,]    4    1    2    3    5    6
#> [351,]    4    1    2    6    3    5
#> [352,]    4    1    2    6    5    3
#> [353,]    4    1    2    5    6    3
#> [354,]    4    1    2    5    3    6
#> [355,]    4    1    5    3    2    6
#> [356,]    4    1    5    3    6    2
#> [357,]    4    1    5    2    3    6
#> [358,]    4    1    5    2    6    3
#> [359,]    4    1    5    6    2    3
#> [360,]    4    1    5    6    3    2
#> [361,]    3    5    4    6    2    1
#> [362,]    3    5    4    6    1    2
#> [363,]    3    5    4    2    6    1
#> [364,]    3    5    4    2    1    6
#> [365,]    3    5    4    1    2    6
#> [366,]    3    5    4    1    6    2
#> [367,]    3    5    6    4    2    1
#> [368,]    3    5    6    4    1    2
#> [369,]    3    5    6    2    4    1
#> [370,]    3    5    6    2    1    4
#> [371,]    3    5    6    1    2    4
#> [372,]    3    5    6    1    4    2
#> [373,]    3    5    2    6    4    1
#> [374,]    3    5    2    6    1    4
#> [375,]    3    5    2    4    6    1
#> [376,]    3    5    2    4    1    6
#> [377,]    3    5    2    1    4    6
#> [378,]    3    5    2    1    6    4
#> [379,]    3    5    1    6    2    4
#> [380,]    3    5    1    6    4    2
#> [381,]    3    5    1    2    6    4
#> [382,]    3    5    1    2    4    6
#> [383,]    3    5    1    4    2    6
#> [384,]    3    5    1    4    6    2
#> [385,]    3    4    5    6    2    1
#> [386,]    3    4    5    6    1    2
#> [387,]    3    4    5    2    6    1
#> [388,]    3    4    5    2    1    6
#> [389,]    3    4    5    1    2    6
#> [390,]    3    4    5    1    6    2
#> [391,]    3    4    6    5    2    1
#> [392,]    3    4    6    5    1    2
#> [393,]    3    4    6    2    5    1
#> [394,]    3    4    6    2    1    5
#> [395,]    3    4    6    1    2    5
#> [396,]    3    4    6    1    5    2
#> [397,]    3    4    2    6    5    1
#> [398,]    3    4    2    6    1    5
#> [399,]    3    4    2    5    6    1
#> [400,]    3    4    2    5    1    6
#> [401,]    3    4    2    1    5    6
#> [402,]    3    4    2    1    6    5
#> [403,]    3    4    1    6    2    5
#> [404,]    3    4    1    6    5    2
#> [405,]    3    4    1    2    6    5
#> [406,]    3    4    1    2    5    6
#> [407,]    3    4    1    5    2    6
#> [408,]    3    4    1    5    6    2
#> [409,]    3    6    4    5    2    1
#> [410,]    3    6    4    5    1    2
#> [411,]    3    6    4    2    5    1
#> [412,]    3    6    4    2    1    5
#> [413,]    3    6    4    1    2    5
#> [414,]    3    6    4    1    5    2
#> [415,]    3    6    5    4    2    1
#> [416,]    3    6    5    4    1    2
#> [417,]    3    6    5    2    4    1
#> [418,]    3    6    5    2    1    4
#> [419,]    3    6    5    1    2    4
#> [420,]    3    6    5    1    4    2
#> [421,]    3    6    2    5    4    1
#> [422,]    3    6    2    5    1    4
#> [423,]    3    6    2    4    5    1
#> [424,]    3    6    2    4    1    5
#> [425,]    3    6    2    1    4    5
#> [426,]    3    6    2    1    5    4
#> [427,]    3    6    1    5    2    4
#> [428,]    3    6    1    5    4    2
#> [429,]    3    6    1    2    5    4
#> [430,]    3    6    1    2    4    5
#> [431,]    3    6    1    4    2    5
#> [432,]    3    6    1    4    5    2
#> [433,]    3    2    4    6    5    1
#> [434,]    3    2    4    6    1    5
#> [435,]    3    2    4    5    6    1
#> [436,]    3    2    4    5    1    6
#> [437,]    3    2    4    1    5    6
#> [438,]    3    2    4    1    6    5
#> [439,]    3    2    6    4    5    1
#> [440,]    3    2    6    4    1    5
#> [441,]    3    2    6    5    4    1
#> [442,]    3    2    6    5    1    4
#> [443,]    3    2    6    1    5    4
#> [444,]    3    2    6    1    4    5
#> [445,]    3    2    5    6    4    1
#> [446,]    3    2    5    6    1    4
#> [447,]    3    2    5    4    6    1
#> [448,]    3    2    5    4    1    6
#> [449,]    3    2    5    1    4    6
#> [450,]    3    2    5    1    6    4
#> [451,]    3    2    1    6    5    4
#> [452,]    3    2    1    6    4    5
#> [453,]    3    2    1    5    6    4
#> [454,]    3    2    1    5    4    6
#> [455,]    3    2    1    4    5    6
#> [456,]    3    2    1    4    6    5
#> [457,]    3    1    4    6    2    5
#> [458,]    3    1    4    6    5    2
#> [459,]    3    1    4    2    6    5
#> [460,]    3    1    4    2    5    6
#> [461,]    3    1    4    5    2    6
#> [462,]    3    1    4    5    6    2
#> [463,]    3    1    6    4    2    5
#> [464,]    3    1    6    4    5    2
#> [465,]    3    1    6    2    4    5
#> [466,]    3    1    6    2    5    4
#> [467,]    3    1    6    5    2    4
#> [468,]    3    1    6    5    4    2
#> [469,]    3    1    2    6    4    5
#> [470,]    3    1    2    6    5    4
#> [471,]    3    1    2    4    6    5
#> [472,]    3    1    2    4    5    6
#> [473,]    3    1    2    5    4    6
#> [474,]    3    1    2    5    6    4
#> [475,]    3    1    5    6    2    4
#> [476,]    3    1    5    6    4    2
#> [477,]    3    1    5    2    6    4
#> [478,]    3    1    5    2    4    6
#> [479,]    3    1    5    4    2    6
#> [480,]    3    1    5    4    6    2
#> [481,]    2    5    4    3    6    1
#> [482,]    2    5    4    3    1    6
#> [483,]    2    5    4    6    3    1
#> [484,]    2    5    4    6    1    3
#> [485,]    2    5    4    1    6    3
#> [486,]    2    5    4    1    3    6
#> [487,]    2    5    3    4    6    1
#> [488,]    2    5    3    4    1    6
#> [489,]    2    5    3    6    4    1
#> [490,]    2    5    3    6    1    4
#> [491,]    2    5    3    1    6    4
#> [492,]    2    5    3    1    4    6
#> [493,]    2    5    6    3    4    1
#> [494,]    2    5    6    3    1    4
#> [495,]    2    5    6    4    3    1
#> [496,]    2    5    6    4    1    3
#> [497,]    2    5    6    1    4    3
#> [498,]    2    5    6    1    3    4
#> [499,]    2    5    1    3    6    4
#> [500,]    2    5    1    3    4    6
#> [501,]    2    5    1    6    3    4
#> [502,]    2    5    1    6    4    3
#> [503,]    2    5    1    4    6    3
#> [504,]    2    5    1    4    3    6
#> [505,]    2    4    5    3    6    1
#> [506,]    2    4    5    3    1    6
#> [507,]    2    4    5    6    3    1
#> [508,]    2    4    5    6    1    3
#> [509,]    2    4    5    1    6    3
#> [510,]    2    4    5    1    3    6
#> [511,]    2    4    3    5    6    1
#> [512,]    2    4    3    5    1    6
#> [513,]    2    4    3    6    5    1
#> [514,]    2    4    3    6    1    5
#> [515,]    2    4    3    1    6    5
#> [516,]    2    4    3    1    5    6
#> [517,]    2    4    6    3    5    1
#> [518,]    2    4    6    3    1    5
#> [519,]    2    4    6    5    3    1
#> [520,]    2    4    6    5    1    3
#> [521,]    2    4    6    1    5    3
#> [522,]    2    4    6    1    3    5
#> [523,]    2    4    1    3    6    5
#> [524,]    2    4    1    3    5    6
#> [525,]    2    4    1    6    3    5
#> [526,]    2    4    1    6    5    3
#> [527,]    2    4    1    5    6    3
#> [528,]    2    4    1    5    3    6
#> [529,]    2    3    4    5    6    1
#> [530,]    2    3    4    5    1    6
#> [531,]    2    3    4    6    5    1
#> [532,]    2    3    4    6    1    5
#> [533,]    2    3    4    1    6    5
#> [534,]    2    3    4    1    5    6
#> [535,]    2    3    5    4    6    1
#> [536,]    2    3    5    4    1    6
#> [537,]    2    3    5    6    4    1
#> [538,]    2    3    5    6    1    4
#> [539,]    2    3    5    1    6    4
#> [540,]    2    3    5    1    4    6
#> [541,]    2    3    6    5    4    1
#> [542,]    2    3    6    5    1    4
#> [543,]    2    3    6    4    5    1
#> [544,]    2    3    6    4    1    5
#> [545,]    2    3    6    1    4    5
#> [546,]    2    3    6    1    5    4
#> [547,]    2    3    1    5    6    4
#> [548,]    2    3    1    5    4    6
#> [549,]    2    3    1    6    5    4
#> [550,]    2    3    1    6    4    5
#> [551,]    2    3    1    4    6    5
#> [552,]    2    3    1    4    5    6
#> [553,]    2    6    4    3    5    1
#> [554,]    2    6    4    3    1    5
#> [555,]    2    6    4    5    3    1
#> [556,]    2    6    4    5    1    3
#> [557,]    2    6    4    1    5    3
#> [558,]    2    6    4    1    3    5
#> [559,]    2    6    3    4    5    1
#> [560,]    2    6    3    4    1    5
#> [561,]    2    6    3    5    4    1
#> [562,]    2    6    3    5    1    4
#> [563,]    2    6    3    1    5    4
#> [564,]    2    6    3    1    4    5
#> [565,]    2    6    5    3    4    1
#> [566,]    2    6    5    3    1    4
#> [567,]    2    6    5    4    3    1
#> [568,]    2    6    5    4    1    3
#> [569,]    2    6    5    1    4    3
#> [570,]    2    6    5    1    3    4
#> [571,]    2    6    1    3    5    4
#> [572,]    2    6    1    3    4    5
#> [573,]    2    6    1    5    3    4
#> [574,]    2    6    1    5    4    3
#> [575,]    2    6    1    4    5    3
#> [576,]    2    6    1    4    3    5
#> [577,]    2    1    4    3    6    5
#> [578,]    2    1    4    3    5    6
#> [579,]    2    1    4    6    3    5
#> [580,]    2    1    4    6    5    3
#> [581,]    2    1    4    5    6    3
#> [582,]    2    1    4    5    3    6
#> [583,]    2    1    3    4    6    5
#> [584,]    2    1    3    4    5    6
#> [585,]    2    1    3    6    4    5
#> [586,]    2    1    3    6    5    4
#> [587,]    2    1    3    5    6    4
#> [588,]    2    1    3    5    4    6
#> [589,]    2    1    6    3    4    5
#> [590,]    2    1    6    3    5    4
#> [591,]    2    1    6    4    3    5
#> [592,]    2    1    6    4    5    3
#> [593,]    2    1    6    5    4    3
#> [594,]    2    1    6    5    3    4
#> [595,]    2    1    5    3    6    4
#> [596,]    2    1    5    3    4    6
#> [597,]    2    1    5    6    3    4
#> [598,]    2    1    5    6    4    3
#> [599,]    2    1    5    4    6    3
#> [600,]    2    1    5    4    3    6
#> [601,]    1    5    4    3    2    6
#> [602,]    1    5    4    3    6    2
#> [603,]    1    5    4    2    3    6
#> [604,]    1    5    4    2    6    3
#> [605,]    1    5    4    6    2    3
#> [606,]    1    5    4    6    3    2
#> [607,]    1    5    3    4    2    6
#> [608,]    1    5    3    4    6    2
#> [609,]    1    5    3    2    4    6
#> [610,]    1    5    3    2    6    4
#> [611,]    1    5    3    6    2    4
#> [612,]    1    5    3    6    4    2
#> [613,]    1    5    2    3    4    6
#> [614,]    1    5    2    3    6    4
#> [615,]    1    5    2    4    3    6
#> [616,]    1    5    2    4    6    3
#> [617,]    1    5    2    6    4    3
#> [618,]    1    5    2    6    3    4
#> [619,]    1    5    6    3    2    4
#> [620,]    1    5    6    3    4    2
#> [621,]    1    5    6    2    3    4
#> [622,]    1    5    6    2    4    3
#> [623,]    1    5    6    4    2    3
#> [624,]    1    5    6    4    3    2
#> [625,]    1    4    5    3    2    6
#> [626,]    1    4    5    3    6    2
#> [627,]    1    4    5    2    3    6
#> [628,]    1    4    5    2    6    3
#> [629,]    1    4    5    6    2    3
#> [630,]    1    4    5    6    3    2
#> [631,]    1    4    3    5    2    6
#> [632,]    1    4    3    5    6    2
#> [633,]    1    4    3    2    5    6
#> [634,]    1    4    3    2    6    5
#> [635,]    1    4    3    6    2    5
#> [636,]    1    4    3    6    5    2
#> [637,]    1    4    2    3    5    6
#> [638,]    1    4    2    3    6    5
#> [639,]    1    4    2    5    3    6
#> [640,]    1    4    2    5    6    3
#> [641,]    1    4    2    6    5    3
#> [642,]    1    4    2    6    3    5
#> [643,]    1    4    6    3    2    5
#> [644,]    1    4    6    3    5    2
#> [645,]    1    4    6    2    3    5
#> [646,]    1    4    6    2    5    3
#> [647,]    1    4    6    5    2    3
#> [648,]    1    4    6    5    3    2
#> [649,]    1    3    4    5    2    6
#> [650,]    1    3    4    5    6    2
#> [651,]    1    3    4    2    5    6
#> [652,]    1    3    4    2    6    5
#> [653,]    1    3    4    6    2    5
#> [654,]    1    3    4    6    5    2
#> [655,]    1    3    5    4    2    6
#> [656,]    1    3    5    4    6    2
#> [657,]    1    3    5    2    4    6
#> [658,]    1    3    5    2    6    4
#> [659,]    1    3    5    6    2    4
#> [660,]    1    3    5    6    4    2
#> [661,]    1    3    2    5    4    6
#> [662,]    1    3    2    5    6    4
#> [663,]    1    3    2    4    5    6
#> [664,]    1    3    2    4    6    5
#> [665,]    1    3    2    6    4    5
#> [666,]    1    3    2    6    5    4
#> [667,]    1    3    6    5    2    4
#> [668,]    1    3    6    5    4    2
#> [669,]    1    3    6    2    5    4
#> [670,]    1    3    6    2    4    5
#> [671,]    1    3    6    4    2    5
#> [672,]    1    3    6    4    5    2
#> [673,]    1    2    4    3    5    6
#> [674,]    1    2    4    3    6    5
#> [675,]    1    2    4    5    3    6
#> [676,]    1    2    4    5    6    3
#> [677,]    1    2    4    6    5    3
#> [678,]    1    2    4    6    3    5
#> [679,]    1    2    3    4    5    6
#> [680,]    1    2    3    4    6    5
#> [681,]    1    2    3    5    4    6
#> [682,]    1    2    3    5    6    4
#> [683,]    1    2    3    6    5    4
#> [684,]    1    2    3    6    4    5
#> [685,]    1    2    5    3    4    6
#> [686,]    1    2    5    3    6    4
#> [687,]    1    2    5    4    3    6
#> [688,]    1    2    5    4    6    3
#> [689,]    1    2    5    6    4    3
#> [690,]    1    2    5    6    3    4
#> [691,]    1    2    6    3    5    4
#> [692,]    1    2    6    3    4    5
#> [693,]    1    2    6    5    3    4
#> [694,]    1    2    6    5    4    3
#> [695,]    1    2    6    4    5    3
#> [696,]    1    2    6    4    3    5
#> [697,]    1    6    4    3    2    5
#> [698,]    1    6    4    3    5    2
#> [699,]    1    6    4    2    3    5
#> [700,]    1    6    4    2    5    3
#> [701,]    1    6    4    5    2    3
#> [702,]    1    6    4    5    3    2
#> [703,]    1    6    3    4    2    5
#> [704,]    1    6    3    4    5    2
#> [705,]    1    6    3    2    4    5
#> [706,]    1    6    3    2    5    4
#> [707,]    1    6    3    5    2    4
#> [708,]    1    6    3    5    4    2
#> [709,]    1    6    2    3    4    5
#> [710,]    1    6    2    3    5    4
#> [711,]    1    6    2    4    3    5
#> [712,]    1    6    2    4    5    3
#> [713,]    1    6    2    5    4    3
#> [714,]    1    6    2    5    3    4
#> [715,]    1    6    5    3    2    4
#> [716,]    1    6    5    3    4    2
#> [717,]    1    6    5    2    3    4
#> [718,]    1    6    5    2    4    3
#> [719,]    1    6    5    4    2    3
#> [720,]    1    6    5    4    3    2

perms(c(1, exp(1), pi))

#>          [,1]     [,2]     [,3]
#> [1,] 3.141593 2.718282 1.000000
#> [2,] 3.141593 1.000000 2.718282
#> [3,] 2.718282 3.141593 1.000000
#> [4,] 2.718282 1.000000 3.141593
#> [5,] 1.000000 2.718282 3.141593
#> [6,] 1.000000 3.141593 2.718282