piecewise {pracma}	R Documentation

Piecewise Linear Function

Description

Compute zeros and area of a piecewise linear function.

Usage

piecewise(x, y, abs = FALSE)

Arguments

x, y
x- and y-coordinates of points defining the piecewise linear function

abs
logical; shall the integral or the total area between the x-axis and the function be calculated

Details

Compute zeros and integral resp. area of a piecewise linear function given by points with x and y as coordinates.

Value

Returns a list with the integral or area as first element and the vector as all zeroes as second.

See Also

trapz

Examples

x <- c(0,  2, 3,  4, 5)
y <- c(2, -2, 0, -2, 0)
piecewise(x, y)
piecewise(x, y, abs=TRUE)

[Package pracma version 2.3.8 Index]