polyval, polyvalm {pracma}	R Documentation

Evaluating a Polynomial

Description

Evaluate polynomial on vector or matrix.

Usage

  polyval(p, x)

  polyvalm(p, A)

Arguments

p
vector representing a polynomial.

x
vector of values where to evaluate the polynomial.

A
matrix; needs to be square.

Details

polyval valuates the polynomial given by p at the values specified by the elements of x. If x is a matrix, the polynomial will be evaluated at each element and a matrix returned.

polyvalm will evaluate the polynomial in the matrix sense, i.e., matrix multiplication is used instead of element by element multiplication as used in 'polyval'. The argument matrix A must be a square matrix.

Value

Vector of values, resp. a matrix.

See Also

poly, roots

Examples

  # Evaluate 3 x^2 + 2 x + 1 at x = 5, 7, and 9
  p = c(3, 2, 1);
  polyval(p, c(5, 7, 9))    # 86  162  262

  # Apply the characteristic polynomial to its matrix
  A <- pascal(4)
  p <- pracma::Poly(A)      # characteristic polynomial of A
  polyvalm(p, A)            # almost zero 4x4-matrix

[Package pracma version 2.3.8 Index]