quiver {pracma}	R Documentation

Quiver or Velocity Plot

Description

A quiver plot displays velocity vectors as arrows with components (u,v) at the points (x,y).

Usage

    quiver(x, y, u, v,
           scale = 0.05, angle = 10, length = 0.1, ...)

Arguments

x, y
x,y-coordinates of start points of the arrows.

u, v
x,y-coordinates of start points.

scale
scales the length of the arrows.

angle
angle between shaft and edge of the arrows.

length
length of the arrow edges.

...
more options presented to the arrows primitive.

Details

The matrices x, y, u, v must all be the same size and contain corresponding position and velocity components. However, x and y can also be vectors.

Value

Opens a graph window and plots the velocity vectors.

See Also

vectorfield, arrows

[Package pracma version 2.3.8 Index]