Create Random Matrices

Create random matrices or random points in a unit circle (Matlab style).

rand(n = 1, m = n)
randn(n = 1, m = n)
randi(imax, n = 1, m = n)
randsample(n, k, w = NULL, replacement = FALSE)

rands(n = 1, N = 1, r = 1)
randp(n = 1, r = 1)

Arguments

n, m
integers specifying the size of the matrix

imax
integer or pair of integers

k
number of elements to return.

w
weight vector, used for discrete probabilities.

replacement
logical; sampling with or without replacement.

N
dimension of a shere, N=1 for the unit circle

r
radius of circle, default 1.

Details

rand(), randn(), randi() create random matrices of size n x m, where the default is square matrices if m is missing.

rand() uses the uniform distribution on ]0, 1[, while randn() uses the normal distribution with mean 0 and standard deviation 1.

randi() generates integers between imax[1] and imax[2] resp. 1 and imax, if imax is a scalar.

randsample() samples k elements from 1:n, with or without replacement, or returns a weighted sample (with replacement), using the weight vector w for probabilities.

rands() generates uniformly random points on an N-sphere in the N+1-dimensional space. To generate uniformly random points in the N-dim. unit cube, take points in S^{N-1} und multiply with unif(n)^(1/(N-1)).

randp() generates uniformly random points in the unit circle (or in a circle of radius r).

Value

Matrices of size nxm resp. a vector of length n.

randp() returns a pair of values representing a point in the circle, or a matrix of size (n,2). rands() returns a matrix of size (n, N+1) with all rows being vectors of length 1.

Note

The Matlab style of setting a seed is not available; use R style set.seed(...).

References

Knuth, D. (1981). The Art of Computer programming; Vol. 2: Seminumerical Algorithms; Chapt. 3: Random Numbers. Addison-Wesley, Reading.

See also

set.seed

Examples

rand(3)

#>           [,1]         [,2]      [,3]
#> [1,] 0.3297161 0.7133512162 0.2816799
#> [2,] 0.7779866 0.2095251395 0.1012531
#> [3,] 0.1983183 0.0001953258 0.8735681

randn(1, 5)

#>            [,1]      [,2]      [,3]        [,4]      [,5]
#> [1,] -0.8723349 0.9240631 0.5164306 -0.04357112 0.4573032

randi(c(1,6), 1, 10)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    6    3    3    1    2    3    5    1    4     1

randsample(10, 5, replacement = TRUE, w = c(0,0,0, 1, 1, 1, 1, 0,0,0))

#> [1] 7 5 7 4 4

P <- rands(1000, N = 1, r = 2)
U <- randp(1000, 2)
# NOT RUN {
plot(U[, 1], U[, 2], pch = "+", asp = 1)
points(P, pch = ".")
# }

#-- v is 2 independent normally distributed elements
# u <- randp(1); r <- t(u) %*% u
# v <- sqrt(-2 * log(r)/r) * u

n <- 5000; U <- randp(n)
R <- apply(U*U, 1, sum)
P <- sqrt(-2 * log(R)/R) * U  # rnorm(2*n)
# NOT RUN {
hist(c(P))
# }