randperm {pracma}	R Documentation

Random Permutation

Description

Generates a random permutation.

Usage

randperm(a, k)

Arguments

a
integer or numeric vector of some length n.

k
integer, smaller as a or length(a).

Details

Generates one random permutation of k of the elements a, if a is a vector, or of 1:a if a is a single integer.

Value

Vector of permuted elements of a or 1:a.

Note

This behavior is different from Matlab/Octave, but does better correspond with the behavior of the perms() function.

See Also

perms

Examples

randperm(1:6, 3)
randperm(6, 6)
randperm(11:20, 5)
randperm(seq(2, 10, by=2))

[Package pracma version 2.3.8 Index]