Reshape {pracma}	R Documentation

Reshape Matrix

Description

Reshape matrix or vector.

Usage

Reshape(a, n, m)

Arguments

a
matrix or vector

n, m
size of the result

Details

Reshape(a, n, m) returns the n-by-m matrix whose elements are taken column-wise from a.

An error results if a does not have n*m elements. If m is missing, it will be calculated from n and the size of a.

Value

Returns matrix (or array) of the requested size containing the elements of a.

Examples

a <- matrix(1:12, nrow=4, ncol=3)
Reshape(a, 6, 2)
Reshape(a, 6)     # the same
Reshape(a, 3, 4)

[Package pracma version 2.3.8 Index]