Accuracy Measures

Calculates different accuracy measures, most prominently RMSE.

rmserr(x, y, summary = FALSE)

Arguments

x, y
two vectors of real numbers

summary
logical; should a summary be printed to the screen?

Details

Calculates six different measures of accuracy for two given vectors or sequences of real numbers:

MAE 	 Mean Absolute Error

MSE 	 Mean Squared Error

RMSE 	 Root Mean Squared Error

MAPE 	 Mean Absolute Percentage Error

LMSE 	 Normalized Mean Squared Error

Value

Returns a list with different accuracy measures.

References

Gentle, J. E. (2009). Computational Statistics, section 10.3. Springer Science+Business Media LCC, New York.

Note

Often used in Data Mining for predicting the accuracy of predictions.

Examples

x <- rep(1, 10)
y <- rnorm(10, 1, 0.1)
rmserr(x, y, summary = TRUE)

#> -- Error Terms --------------------------------------------------
#>  MAE:   0.0761   	- mean absolute error (in range [ 1 1 ])
#>  MSE:   0.0089   	- mean squared error (the variance?!)
#>  RMSE:  0.0944   	- root mean squared error (std. dev.)
#>  MAPE:  0.0761   	- mean absolute percentage error
#>  LMSE:    Inf   	- normalized mean squared error
#>  rSTD:  0.0944   	- relative standard deviation ( 1 )
#> -----------------------------------------------------------------