Rosser Matrix

Generate the Rosser matrix.

rosser()

Details

This is a classic symmetric eigenvalue test problem. It has a double eigenvalue, three nearly equal eigenvalues, dominant eigenvalues of opposite sign, a zero eigenvalue, and a small, nonzero eigenvalue.

Value

matrix of size 8 x 8

See also

wilkinson

Examples

rosser()

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,]  611  196 -192  407   -8  -52  -49   29
#> [2,]  196  899  113 -192  -71  -43   -8  -44
#> [3,] -192  113  899  196   61   49    8   52
#> [4,]  407 -192  196  611    8   44   59  -23
#> [5,]   -8  -71   61    8  411 -599  208  208
#> [6,]  -52  -43   49   44 -599  411  208  208
#> [7,]  -49   -8    8   59  208  208   99 -911
#> [8,]   29  -44   52  -23  208  208 -911   99