size {pracma}	R Documentation

Size of Matrix

Description

Provides the dimensions of x.

Usage

size(x, k)

Arguments

x
vector, matrix, or array

k
integer specifying a particular dimension

Details

Returns the number of dimensions as length(x).

Vector will be treated as a single row matrix.

Value

vector containing the dimensions of x, or the k-th dimension if k is not missing.

Note

The result will differ from Matlab when x is a character vector.

See Also

dim

Examples

size(1:8)
size(matrix(1:8, 2, 4))		# 2 4
size(matrix(1:8, 2, 4), 2)	# 4
size(matrix(1:8, 2, 4), 3)	# 1

[Package pracma version 2.3.8 Index]