sortrows {pracma}	R Documentation

Sort Rows of a Matrix (Matlab Style)

Description

Sort rows of a matrix according to values in a column.

Usage

sortrows(A, k = 1)

Arguments

A
numeric matrix.

k
number of column to sort the matrix accordingly.

Details

sortrows(A, k) sorts the rows of the matrix A such that column k is increasingly sorted.

Value

Returns the sorted matrix.

See Also

sort

Examples

A <- magic(5)
sortrows(A)
sortrows(A, k = 2)

[Package pracma version 2.3.8 Index]