strcmp {pracma}	R Documentation

String Comparison

Description

Compare two strings or character vectors for equality.

Usage

strcmp(s1, s2)
strcmpi(s1, s2)

Arguments

s1, s2
character strings or vectors

Details

For strcmp comparisons are case-sensitive, while for strcmpi the are case-insensitive. Leading and trailing blanks do count.

Value

logical, i.e. TRUE if s1 and s2 have the same length as character vectors and all elements are equal as character strings, else FALSE.

See Also

strcat

Examples

strcmp(c("yes", "no"), c("yes", "no"))
strcmpi(c("yes", "no"), c("Yes", "No"))

[Package pracma version 2.3.8 Index]