strTrim {pracma}	R Documentation

Remove leading and trailing white space.

Description

Removes leading and trailing white space from a string.

Usage

strTrim(s)
deblank(s)

Arguments

s
character string or character vector

Details

strTrim removes leading and trailing white space from a string or from all strings in a character vector.

deblank removes trailing white space only from a string or from all strings in a character vector.

Value

A character string or character vector with (leading and) trailing white space.

See Also

strjust

Examples

s <- c("  abc", "abc   ", " abc ", " a b c ", "abc", "a b c")
strTrim(s)
deblank(s)

[Package pracma version 2.3.8 Index]