tic,toc {pracma}	R Documentation

MATLAB timer functions

Description

Provides stopwatch timer. Function tic starts the timer and toc updates the elapsed time since the timer was started.

Usage

tic(gcFirst=FALSE)
toc(echo=TRUE)

Arguments

gcFirst
logical scalar. If TRUE, perform garbage collection prior to starting stopwatch

echo
logical scalar. If TRUE, print elapsed time to screen

Details

Provides analog to system.time. Function toc can be invoked multiple times in a row.

Value

toc invisibly returns the elapsed time as a named scalar (vector).

Author(s)

P. Roebuck proebuck@mdanderson.org

Examples

tic()
for(i in 1:100) mad(runif(1000))	# kill time
toc()

[Package pracma version 2.3.8 Index]