titanium {pracma}	R Documentation

Titanium Test Data

Description

The Titanium data set describes measurements of a certain property of titanium as a function of temperature.

Usage

data(titanium)

Format

The format is:
Two columns called ‘x’ and ‘y’, the first being the temperature.

Details

These data have become a standard test for data fitting since they are hard to fit by classical techniques and have a significant amount of noise.

Source

Boor, C. de, and J. R. Rice (1968). Least squares cubic spline approximation II – Variable knots, CSD TR 21, Comp.Sci., Purdue Univ.

Examples

## Not run: 
data(titanium)
plot(titanium)
grid()
## End(Not run)

[Package pracma version 2.3.8 Index]