Toeplitz {pracma}	R Documentation

Toeplitz Matrix

Description

Generate Toeplitz matrix from column and row vector.

Usage

Toeplitz(a, b)

Arguments

a
vector that will be the first column

b
vector that if present will form the first row.

Details

Toeplitz(a, b) returns a (non-symmetric) Toeplitz matrix whose first column is a and whose first row is b. The following rows are shifted to the left.

If the first element of b differs from the last element of a it is overwritten by this one (and a warning sent).

Value

Matrix of size (length(a), length(b)).

Note

stats::Toeplitz does not allow to specify the row vector, that is returns only the symmetric Toeplitz matrix.

See Also

hankel

Examples

Toeplitz(c(1, 2, 3, 4, 5))
Toeplitz(c(1, 2, 3, 4, 5), c(1.5, 2.5, 3.5, 4.5, 5.5))

[Package pracma version 2.3.8 Index]