tri {pracma}	R Documentation

Triangular Matrices (Matlab Style)

Description

Extract lower or upper triangular part of a matrix.

Usage

tril(M, k = 0)
triu(M, k = 0)

Arguments

M
numeric matrix.

k
integer, indicating a secondary diagonal.

Details

tril
Returns the elements on and below the kth diagonal of X, where k = 0 is the main diagonal, k > 0 is above the main diagonal, and k < 0 is below the main diagonal.

triu
Returns the elements on and above the kth diagonal of X, where k = 0 is the main diagonal, k > 0 is above the main diagonal, and k < 0 is below the main diagonal.

Value

Matrix the same size as the input matrix.

Note

For k==0 it is simply an application of the R functions lower.tri resp. upper.tri.

See Also

Diag

Examples

tril(ones(4,4), +1)
#    1  1  0  0
#    1  1  1  0
#    1  1  1  1
#    1  1  1  1

triu(ones(4,4), -1)
#    1  1  1  1
#    1  1  1  1
#    0  1  1  1
#    0  0  1  1

[Package pracma version 2.3.8 Index]