cross {pracma}	R Documentation

Vector Cross Product

Description

Vector or cross product

Usage

  cross(x, y)

Arguments

x
numeric vector or matrix

y
numeric vector or matrix

Details

Computes the cross (or: vector) product of vectors in 3 dimensions. In case of matrices it takes the first dimension of length 3 and computes the cross product between corresponding columns or rows.

The more general cross product of n-1 vectors in n-dimensional space is realized as crossn.

Value

3-dim. vector if x and < are vectors, a matrix of 3-dim. vectors if x and y are matrices themselves.

See Also

dot, crossn

Examples

  cross(c(1, 2, 3), c(4, 5, 6))  # -3  6 -3

[Package pracma version 2.3.8 Index]