Generate Permutations

Generates all permutations of a vector a.

perms(a)

Arguments

a
numeric vector of some length n

Details

If a is a vector of length n, generate all permutations of the elements in a as a matrix of size n! x n where each row represents one permutation.

A matrix will be expanded as vector.

Value

matrix of permutations of the elements of a

Note

Not feasible for length(a) > 10.

See also

randperm

Examples

perms(6)

#>      [,1]
#> [1,]    6

perms(1:6)

#>        [,1] [,2] [,3] [,4] [,5] [,6]
#>   [1,]    6    5    4    3    2    1
#>   [2,]    6    5    4    3    1    2
#>   [3,]    6    5    4    2    3    1
#>   [4,]    6    5    4    2    1    3
#>   [5,]    6    5    4    1    2    3
#>   [6,]    6    5    4    1    3    2
#>   [7,]    6    5    3    4    2    1
#>   [8,]    6    5    3    4    1    2
#>   [9,]    6    5    3    2    4    1
#>  [10,]    6    5    3    2    1    4
#>  [11,]    6    5    3    1    2    4
#>  [12,]    6    5    3    1    4    2
#>  [13,]    6    5    2    3    4    1
#>  [14,]    6    5    2    3    1    4
#>  [15,]    6    5    2    4    3    1
#>  [16,]    6    5    2    4    1    3
#>  [17,]    6    5    2    1    4    3
#>  [18,]    6    5    2    1    3    4
#>  [19,]    6    5    1    3    2    4
#>  [20,]    6    5    1    3    4    2
#>  [21,]    6    5    1    2    3    4
#>  [22,]    6    5    1    2    4    3
#>  [23,]    6    5    1    4    2    3
#>  [24,]    6    5    1    4    3    2
#>  [25,]    6    4    5    3    2    1
#>  [26,]    6    4    5    3    1    2
#>  [27,]    6    4    5    2    3    1
#>  [28,]    6    4    5    2    1    3
#>  [29,]    6    4    5    1    2    3
#>  [30,]    6    4    5    1    3    2
#>  [31,]    6    4    3    5    2    1
#>  [32,]    6    4    3    5    1    2
#>  [33,]    6    4    3    2    5    1
#>  [34,]    6    4    3    2    1    5
#>  [35,]    6    4    3    1    2    5
#>  [36,]    6    4    3    1    5    2
#>  [37,]    6    4    2    3    5    1
#>  [38,]    6    4    2    3    1    5
#>  [39,]    6    4    2    5    3    1
#>  [40,]    6    4    2    5    1    3
#>  [41,]    6    4    2    1    5    3
#>  [42,]    6    4    2    1    3    5
#>  [43,]    6    4    1    3    2    5
#>  [44,]    6    4    1    3    5    2
#>  [45,]    6    4    1    2    3    5
#>  [46,]    6    4    1    2    5    3
#>  [47,]    6    4    1    5    2    3
#>  [48,]    6    4    1    5    3    2
#>  [49,]    6    3    4    5    2    1
#>  [50,]    6    3    4    5    1    2
#>  [51,]    6    3    4    2    5    1
#>  [52,]    6    3    4    2    1    5
#>  [53,]    6    3    4    1    2    5
#>  [54,]    6    3    4    1    5    2
#>  [55,]    6    3    5    4    2    1
#>  [56,]    6    3    5    4    1    2
#>  [57,]    6    3    5    2    4    1
#>  [58,]    6    3    5    2    1    4
#>  [59,]    6    3    5    1    2    4
#>  [60,]    6    3    5    1    4    2
#>  [61,]    6    3    2    5    4    1
#>  [62,]    6    3    2    5    1    4
#>  [63,]    6    3    2    4    5    1
#>  [64,]    6    3    2    4    1    5
#>  [65,]    6    3    2    1    4    5
#>  [66,]    6    3    2    1    5    4
#>  [67,]    6    3    1    5    2    4
#>  [68,]    6    3    1    5    4    2
#>  [69,]    6    3    1    2    5    4
#>  [70,]    6    3    1    2    4    5
#>  [71,]    6    3    1    4    2    5
#>  [72,]    6    3    1    4    5    2
#>  [73,]    6    2    4    3    5    1
#>  [74,]    6    2    4    3    1    5
#>  [75,]    6    2    4    5    3    1
#>  [76,]    6    2    4    5    1    3
#>  [77,]    6    2    4    1    5    3
#>  [78,]    6    2    4    1    3    5
#>  [79,]    6    2    3    4    5    1
#>  [80,]    6    2    3    4    1    5
#>  [81,]    6    2    3    5    4    1
#>  [82,]    6    2    3    5    1    4
#>  [83,]    6    2    3    1    5    4
#>  [84,]    6    2    3    1    4    5
#>  [85,]    6    2    5    3    4    1
#>  [86,]    6    2    5    3    1    4
#>  [87,]    6    2    5    4    3    1
#>  [88,]    6    2    5    4    1    3
#>  [89,]    6    2    5    1    4    3
#>  [90,]    6    2    5    1    3    4
#>  [91,]    6    2    1    3    5    4
#>  [92,]    6    2    1    3    4    5
#>  [93,]    6    2    1    5    3    4
#>  [94,]    6    2    1    5    4    3
#>  [95,]    6    2    1    4    5    3
#>  [96,]    6    2    1    4    3    5
#>  [97,]    6    1    4    3    2    5
#>  [98,]    6    1    4    3    5    2
#>  [99,]    6    1    4    2    3    5
#> [100,]    6    1    4    2    5    3
#> [101,]    6    1    4    5    2    3
#> [102,]    6    1    4    5    3    2
#> [103,]    6    1    3    4    2    5
#> [104,]    6    1    3    4    5    2
#> [105,]    6    1    3    2    4    5
#> [106,]    6    1    3    2    5    4
#> [107,]    6    1    3    5    2    4
#> [108,]    6    1    3    5    4    2
#> [109,]    6    1    2    3    4    5
#> [110,]    6    1    2    3    5    4
#> [111,]    6    1    2    4    3    5
#> [112,]    6    1    2    4    5    3
#> [113,]    6    1    2    5    4    3
#> [114,]    6    1    2    5    3    4
#> [115,]    6    1    5    3    2    4
#> [116,]    6    1    5    3    4    2
#> [117,]    6    1    5    2    3    4
#> [118,]    6    1    5    2    4    3
#> [119,]    6    1    5    4    2    3
#> [120,]    6    1    5    4    3    2
#> [121,]    5    6    4    3    2    1
#> [122,]    5    6    4    3    1    2
#> [123,]    5    6    4    2    3    1
#> [124,]    5    6    4    2    1    3
#> [125,]    5    6    4    1    2    3
#> [126,]    5    6    4    1    3    2
#> [127,]    5    6    3    4    2    1
#> [128,]    5    6    3    4    1    2
#> [129,]    5    6    3    2    4    1
#> [130,]    5    6    3    2    1    4
#> [131,]    5    6    3    1    2    4
#> [132,]    5    6    3    1    4    2
#> [133,]    5    6    2    3    4    1
#> [134,]    5    6    2    3    1    4
#> [135,]    5    6    2    4    3    1
#> [136,]    5    6    2    4    1    3
#> [137,]    5    6    2    1    4    3
#> [138,]    5    6    2    1    3    4
#> [139,]    5    6    1    3    2    4
#> [140,]    5    6    1    3    4    2
#> [141,]    5    6    1    2    3    4
#> [142,]    5    6    1    2    4    3
#> [143,]    5    6    1    4    2    3
#> [144,]    5    6    1    4    3    2
#> [145,]    5    4    6    3    2    1
#> [146,]    5    4    6    3    1    2
#> [147,]    5    4    6    2    3    1
#> [148,]    5    4    6    2    1    3
#> [149,]    5    4    6    1    2    3
#> [150,]    5    4    6    1    3    2
#> [151,]    5    4    3    6    2    1
#> [152,]    5    4    3    6    1    2
#> [153,]    5    4    3    2    6    1
#> [154,]    5    4    3    2    1    6
#> [155,]    5    4    3    1    2    6
#> [156,]    5    4    3    1    6    2
#> [157,]    5    4    2    3    6    1
#> [158,]    5    4    2    3    1    6
#> [159,]    5    4    2    6    3    1
#> [160,]    5    4    2    6    1    3
#> [161,]    5    4    2    1    6    3
#> [162,]    5    4    2    1    3    6
#> [163,]    5    4    1    3    2    6
#> [164,]    5    4    1    3    6    2
#> [165,]    5    4    1    2    3    6
#> [166,]    5    4    1    2    6    3
#>  [ reached getOption("max.print") -- omitted 554 rows ]

perms(c(1, exp(1), pi))

#>          [,1]     [,2]     [,3]
#> [1,] 3.141593 2.718282 1.000000
#> [2,] 3.141593 1.000000 2.718282
#> [3,] 2.718282 3.141593 1.000000
#> [4,] 2.718282 1.000000 3.141593
#> [5,] 1.000000 2.718282 3.141593
#> [6,] 1.000000 3.141593 2.718282