rmserr {pracma}	R Documentation

Accuracy Measures

Description

Calculates different accuracy measures, most prominently RMSE.

Usage

rmserr(x, y, summary = FALSE)

Arguments

x, y
two vectors of real numbers

summary
logical; should a summary be printed to the screen?

Details

Calculates six different measures of accuracy for two given vectors or sequences of real numbers:

MAE 	 Mean Absolute Error

MSE 	 Mean Squared Error

RMSE 	 Root Mean Squared Error

MAPE 	 Mean Absolute Percentage Error

LMSE 	 Normalized Mean Squared Error

rSTD 	 relative Standard Deviation

Value

Returns a list with different accuracy measures.

Note

Often used in Data Mining for predicting the accuracy of predictions.

References

Gentle, J. E. (2009). Computational Statistics, section 10.3. Springer Science+Business Media LCC, New York.

Examples

x <- rep(1, 10)
y <- rnorm(10, 1, 0.1)
rmserr(x, y, summary = TRUE)

[Package pracma version 2.3.8 Index]