strjust {pracma}	R Documentation

Justify character vector

Description

Justify the strings in a character vector.

Usage

strjust(s, justify = c("left", "right", "center"))

Arguments

s
Character vector.

justify
Whether to justify left, right, or centered.

Details

strjust(s) or strjust(s, justify = ``right'') returns a right-justified character vector. All strings have the same length, the length of the longest string in s — but the strings in s have been trimmed before.

strjust(s, justify = ``left'') does the same, with all strings left-justified.

strjust(s, justify = ``centered'') returns all string in s centered. If an odd number of blanks has to be added, one blank more is added to the left than to the right.

Value

A character vector of the same length.

See Also

strTrim

Examples

S <- c("abc", "letters", "1", "2  2")
strjust(S, "left")

[Package pracma version 2.3.8 Index]