trigApprox {pracma}	R Documentation

Trigonometric Approximation

Description

Computes the trigonometric series.

Usage

trigApprox(t, x, m)

Arguments

t
vector of points at which to compute the values of the trigonometric approximation.

x
data from t=0 to t=2*(n-1)*pi/n.

m
degree of trigonometric regression.

Details

Calls trigPoly to get the trigonometric coefficients and then sums the finite series.

Value

Vector of values the same length as t.

Note

TODO: Return an approximating function instead.

See Also

trigPoly

Examples

## Not run: 
##  Example: Gauss' Pallas data (1801)
asc <- seq(0, 330, by = 30)
dec <- c(408, 89, -66, 10, 338, 807, 1238, 1511, 1583, 1462, 1183, 804)
plot(2*pi*asc/360, dec, pch = "+", col = "red", xlim = c(0, 2*pi), ylim = c(-500, 2000),
     xlab = "Ascension [radians]", ylab = "Declination [minutes]",
     main = "Gauss' Pallas Data")
grid()
points(2*pi*asc/360, dec, pch = "o", col = "red")
ts <- seq(0, 2*pi, len = 100)
xs <- trigApprox(ts ,dec, 1)
lines(ts, xs, col = "black")
xs <- trigApprox(ts ,dec, 2)
lines(ts, xs, col = "blue")
legend(3, 0, c("Trig. Regression of degree 1", "Trig. Regression of degree 2",
                  "Astronomical position"), col = c("black", "blue", "red"),
                lty = c(1,1,0), pch = c("", "", "+"))
## End(Not run)

[Package pracma version 2.3.8 Index]